requirements.txt
Django
psycopg2
psycopg2-binary
pillow
tensorflow
sorl-thumbnail